Extract explorer input parsing into its own module
